+++ Realm/Realm/Sync/ProgressNotifications/ProgressMode.cs
@@ -25,15 +25,12 @@ namespace Realms.Sync
     {
         /// <summary>
         /// The callback will be called forever, or until it is unregistered by disposing the subscription token.
-        /// Notifications will always report the latest number of transferred bytes, and the most up-to-date number of
-        /// total transferable bytes.
         /// </summary>
         ReportIndefinitely,
 
         /// <summary>
-        /// The callback will, upon registration, store the total number of bytes to be transferred. When invoked, it will
-        /// always report the most up-to-date number of transferable bytes out of that original number of transferable bytes.
-        /// When the number of transferred bytes reaches or exceeds the number of transferable bytes, the callback will
+        /// The callback will be active until the current transferable bytes are transferred. When invoked, it will always report the
+        /// progress estimate as a percentage related to the initial transferable bytes. When those bytes are transfered, the callback will
         /// be unregistered.
         /// </summary>
         ForCurrentlyOutstandingWork